Create arith_formatter
It is proj 1 from course Scientific Computing with Python.
Here arithematic_arranger is a function that takes inputs as list of strings and arranges them one below another.
Function takes two parameters, one the input numbers and second one in boolean type. If boolean is set to "True" then we can get the answer of the input operator.